Safari-611.1.10/LayoutTests/js/builtins/shielding-typedarray-expected.txt

Test to ensure TypedArray filter is shielded from Array.prototype.push modifications

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array.length === 1 is true
PASS successfullyParsed is true

TEST COMPLETE
